Move tokens to its own file
